requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
requests>=2.31.0
python-dotenv
python-escpos>=3.1
psutil>=5.9.0
